+++ src/Epr.Reprocessor.Exporter.UI.App/Services/RegistrationMaterialService.cs
@@ -158,8 +158,3 @@ public class RegistrationMaterialService(
         }
     }
 }
-
-public record CreateRegistrationMaterialResponseDto
-{
-    public Guid Id { get; set; }
-}
